example/requirements.txt
Django~=3.1
django-germanium==2.2.12
coverage==5.0.4
django-reversion>=2.0.13
responses==0.9.0
celery==5.0.5
freezegun==0.3.12
elasticsearch-dsl==7.4.0
https://github.com/druids/django-celery-extensions/tarball/NewResultWrapper#egg=django-celery-extensions
-e ../
